+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<<<<<<< HEAD
 ## [0.15.0] - 2025-05-15
 MR #16
 
@@ -18,6 +19,8 @@ MR #16
 ### Modified
 - `console_handler` in the logging config to only show WARNING and above
 
+=======
+>>>>>>> main
 ## [0.14.0] - 2025-05-11
 MR #15