+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [0.13.0] - 2025-13-05
 MR #16
 
@@ -18,8 +19,15 @@ MR #16
 - `get_sql_query` from `extractor` to utils folder for global use
 - Moved all queries to utils folder for easier use of queries through the whole project
 
-## [0.12.0] - 2025-05-09
-MR #13
+## [0.14.0] - 2025-05-11
+MR #15
+
+### Added
+- Adaptative predictions using ARIMA model, with custom parameter for each station by finding the optimimal ones
+- Combined time-weighted historical data, recent observations, and ARIMA models
+
+## [0.13.0] - 2025-05-09
+MR #14
 
 ### Added
 - Calculation of the walking trip to reach the nearest station using the graph
@@ -27,6 +35,12 @@ MR #13
 ### Changed
 - `graph_builder.find_optimal_route` to use this new function to calculate the whole journey
 - Transfers duration from 5 minutes to 3
+
+## [0.12.0] - 2025-05-06
+MR #13
+
+### Added
+- Traffic predictions with ARIMA
 
 ## [0.11.0] - 2025-05-05
 MR #12